remove v2_as_genesis, use as the default
