+++ crates/e2e/src/tests/consensus_rpc.rs
@@ -35,10 +35,7 @@ async fn consensus_subscribe_and_query_finalization() {
     let _ = tempo_eyre::install();
 
     let initial_height = 3;
-    let setup = Setup::new()
-        .how_many_signers(1)
-        .t2_time(0)
-        .epoch_length(100);
+    let setup = Setup::new().how_many_signers(1).epoch_length(100);
     let cfg = deterministic::Config::default().with_seed(setup.seed);
 
     let (addr_tx, addr_rx) = oneshot::channel::<(SocketAddr, SocketAddr)>();
@@ -188,7 +185,7 @@ fn get_identity_transition_proof_after_full_dkg() {
 
         // --- First full DKG ---
         execution_runtime
-            .set_next_full_dkg_ceremony(http_url.clone(), first_full_dkg_epoch)
+            .set_next_full_dkg_ceremony_v2(http_url.clone(), first_full_dkg_epoch)
             .await
             .unwrap();
 
@@ -220,7 +217,7 @@ fn get_identity_transition_proof_after_full_dkg() {
 
         // --- Second full DKG ---
         execution_runtime
-            .set_next_full_dkg_ceremony(http_url.clone(), second_full_dkg_epoch)
+            .set_next_full_dkg_ceremony_v2(http_url.clone(), second_full_dkg_epoch)
             .await
             .unwrap();
