+++ crates/e2e/src/tests/dkg/static_transitions.rs
@@ -4,7 +4,7 @@
 //! at genesis.
 use commonware_macros::test_traced;
 
-use crate::{Setup, run};
+use crate::{Setup, run, tests::assert_no_v1};
 
 #[test_traced]
 fn single_validator_can_transition_once() {
@@ -98,6 +98,8 @@ impl AssertStaticTransitions {
         let mut epoch_reached = false;
         let mut dkg_successful = false;
         let _first = run(setup, move |metric, value| {
+            assert_no_v1(metric, value);
+
             if metric.ends_with("_dkg_manager_ceremony_failures_total") {
                 let value = value.parse::<u64>().unwrap();
                 assert_eq!(0, value);
