+++ crates/e2e/src/tests/simple.rs
@@ -1,7 +1,7 @@
 //! Simple tests: just start and build a few blocks.
 use std::time::Duration;
 
-use crate::{Setup, run, tests::v2_at_genesis::assert_no_v1};
+use crate::{Setup, run, tests::assert_no_v1};
 use commonware_macros::test_traced;
 use commonware_p2p::simulated::Link;
 
@@ -12,7 +12,6 @@ fn single_node() {
     let setup = Setup::new()
         .how_many_signers(1)
         .epoch_length(100)
-        .t2_time(0)
         .seed(0);
     let _first = run(setup, |metric, value| {
         assert_no_v1(metric, value);
@@ -29,7 +28,7 @@ fn single_node() {
 fn only_good_links() {
     let _ = tempo_eyre::install();
 
-    let setup = Setup::new().epoch_length(100).t2_time(0).seed(42);
+    let setup = Setup::new().epoch_length(100).seed(42);
     let _first = run(setup, |metric, value| {
         assert_no_v1(metric, value);
         if metric.ends_with("_marshal_processed_height") {
@@ -54,7 +53,6 @@ fn many_bad_links() {
     let setup = Setup::new()
         .seed(42)
         .epoch_length(100)
-        .t2_time(0)
         .linkage(link);
 
     let _first = run(setup, |metric, value| {
@@ -81,7 +79,6 @@ fn reach_height_20_with_a_few_bad_links() {
     let setup = Setup::new()
         .how_many_signers(10)
         .epoch_length(100)
-        .t2_time(0)
         .linkage(link);
 
     run(setup, |metric, value| {
