--- crates/e2e/src/tests/v2_at_genesis/dkg/mod.rs
@@ -1,8 +1,0 @@
-//! Tests on chain DKG and epoch transition
-
-pub(crate) mod common;
-mod dynamic;
-mod fast_sync_after_full_dkg;
-mod full_ceremony;
-mod share_loss;
-mod static_transitions;
